requirements.txt
piper-tts
pygame
customtkinter
sounddevice